+++ core/src/state/mod.rs
@@ -50,7 +50,7 @@ use crate::{
     spec::genesis::{
         genesis_contract_address_four_year, genesis_contract_address_two_year,
     },
-    trace::{AddressPocket, Tracer},
+    trace::{AddressPocket, InternalTransferTracer},
     transaction_pool::SharedTransactionPool,
 };
 
@@ -150,12 +150,15 @@ impl<StateDbStorage: StorageStateTrait> StateTrait
     /// checked out. This function should only be called in post-processing
     /// of a transaction.
     fn settle_collateral_for_all(
-        &mut self, substate: &Substate, account_start_nonce: U256,
-    ) -> DbResult<CollateralCheckResult> {
+        &mut self, substate: &Substate,
+        tracer: &mut dyn InternalTransferTracer, account_start_nonce: U256,
+    ) -> DbResult<CollateralCheckResult>
+    {
         for address in substate.keys_for_collateral_changed().iter() {
             match self.settle_collateral_for_address(
                 address,
                 substate,
+                tracer,
                 account_start_nonce,
             )? {
                 CollateralCheckResult::Valid => {}
@@ -169,13 +172,16 @@ impl<StateDbStorage: StorageStateTrait> StateTrait
     // test cases breaks this assumption, which will be fixed in a separated PR.
     fn collect_and_settle_collateral(
         &mut self, original_sender: &Address, storage_limit: &U256,
-        substate: &mut Substate, account_start_nonce: U256,
+        substate: &mut Substate, tracer: &mut dyn InternalTransferTracer,
+        account_start_nonce: U256,
     ) -> DbResult<CollateralCheckResult>
     {
         self.collect_ownership_changed(substate)?;
-        let res = match self
-            .settle_collateral_for_all(substate, account_start_nonce)?
-        {
+        let res = match self.settle_collateral_for_all(
+            substate,
+            tracer,
+            account_start_nonce,
+        )? {
             CollateralCheckResult::Valid => {
                 self.check_storage_limit(original_sender, storage_limit)?
             }
@@ -1173,7 +1179,7 @@ impl<StateDbStorage: StorageStateTrait> StateGeneric<StateDbStorage> {
     /// Charges or refund storage collateral and update `total_storage_tokens`.
     fn settle_collateral_for_address(
         &mut self, addr: &Address, substate: &dyn SubstateTrait,
-        tracer: &mut dyn Tracer, account_start_nonce: U256,
+        tracer: &mut dyn InternalTransferTracer, account_start_nonce: U256,
     ) -> DbResult<CollateralCheckResult>
     {
         let (inc_collaterals, sub_collaterals) =
@@ -1187,12 +1193,12 @@ impl<StateDbStorage: StorageStateTrait> StateGeneric<StateDbStorage> {
 
         if !sub.is_zero() {
             tracer.prepare_internal_transfer_action(
-                /* from */ AddressPocket::StorageCollateral(addr),
+                /* from */ AddressPocket::StorageCollateral(*addr),
                 /* to */
                 if is_contract {
-                    AddressPocket::SponsorBalanceForStorage(addr)
+                    AddressPocket::SponsorBalanceForStorage(*addr)
                 } else {
-                    AddressPocket::Balance(addr)
+                    AddressPocket::Balance(*addr)
                 },
                 sub,
             );
@@ -1214,11 +1220,11 @@ impl<StateDbStorage: StorageStateTrait> StateGeneric<StateDbStorage> {
             tracer.prepare_internal_transfer_action(
                 /* from */
                 if is_contract {
-                    AddressPocket::SponsorBalanceForStorage(addr)
+                    AddressPocket::SponsorBalanceForStorage(*addr)
                 } else {
-                    AddressPocket::Balance(addr)
+                    AddressPocket::Balance(*addr)
                 },
-                /* to */ AddressPocket::StorageCollateral(addr),
+                /* to */ AddressPocket::StorageCollateral(*addr),
                 sub,
             );
