+++ core/src/state/state_tests.rs
@@ -5,7 +5,7 @@
 use super::{CleanupMode, CollateralCheckResult, State, Substate};
 use crate::{
     spec::genesis::DEV_GENESIS_KEY_PAIR,
-    test_helpers::get_state_for_genesis_write, vm::Spec,
+    test_helpers::get_state_for_genesis_write, trace::NoopTracer, vm::Spec,
 };
 use cfx_parameters::{consensus::ONE_CFX_IN_DRIP, staking::*};
 use cfx_state::{
@@ -268,6 +268,7 @@ fn checkpoint_from_empty_get_storage_at() {
                 &a,
                 &U256::MAX,
                 &mut substates.last_mut().unwrap(),
+                &mut NoopTracer,
                 Spec::new_spec_for_test().account_start_nonce
             )
             .unwrap(),
@@ -334,6 +335,7 @@ fn checkpoint_from_empty_get_storage_at() {
                 &a,
                 &U256::MAX,
                 &mut substates.last_mut().unwrap(),
+                &mut NoopTracer,
                 Spec::new_spec_for_test().account_start_nonce
             )
             .unwrap(),
@@ -386,6 +388,7 @@ fn checkpoint_from_empty_get_storage_at() {
                 &a,
                 &U256::MAX,
                 &mut substates.last_mut().unwrap(),
+                &mut NoopTracer,
                 Spec::new_spec_for_test().account_start_nonce
             )
             .unwrap(),
@@ -462,6 +465,7 @@ fn checkpoint_get_storage_at() {
                 &a,
                 &U256::MAX,
                 &mut substates.last_mut().unwrap(),
+                &mut NoopTracer,
                 Spec::new_spec_for_test().account_start_nonce
             )
             .unwrap(),
@@ -609,6 +613,7 @@ fn checkpoint_get_storage_at() {
                 &contract_a,
                 &U256::MAX,
                 &mut substates.last_mut().unwrap(),
+                &mut NoopTracer,
                 Spec::new_spec_for_test().account_start_nonce
             )
             .unwrap(),
@@ -700,6 +705,7 @@ fn checkpoint_get_storage_at() {
                 &contract_a,
                 &U256::MAX,
                 &mut substates.last_mut().unwrap(),
+                &mut NoopTracer,
                 Spec::new_spec_for_test().account_start_nonce
             )
             .unwrap(),
@@ -776,6 +782,7 @@ fn checkpoint_get_storage_at() {
                 &contract_a,
                 &U256::MAX,
                 &mut substates.last_mut().unwrap(),
+                &mut NoopTracer,
                 Spec::new_spec_for_test().account_start_nonce
             )
             .unwrap(),
@@ -997,6 +1004,7 @@ fn create_contract_fail() {
                 &a,
                 &U256::MAX,
                 &mut substate,
+                &mut NoopTracer,
                 Spec::new_spec_for_test().account_start_nonce
             )
             .unwrap(),
@@ -1054,6 +1062,7 @@ fn create_contract_fail_previous_storage() {
                 &a,
                 &U256::MAX,
                 &mut substates.last_mut().unwrap(),
+                &mut NoopTracer,
                 Spec::new_spec_for_test().account_start_nonce
             )
             .unwrap(),
@@ -1218,6 +1227,7 @@ fn test_automatic_collateral_normal_account() {
                 &normal_account,
                 &U256::MAX,
                 &mut substates.last_mut().unwrap(),
+                &mut NoopTracer,
                 Spec::new_spec_for_test().account_start_nonce
             )
             .unwrap(),
@@ -1260,6 +1270,7 @@ fn test_automatic_collateral_normal_account() {
                 &normal_account,
                 &U256::MAX,
                 &mut substates.last_mut().unwrap(),
+                &mut NoopTracer,
                 Spec::new_spec_for_test().account_start_nonce
             )
             .unwrap(),
@@ -1304,6 +1315,7 @@ fn test_automatic_collateral_normal_account() {
                 &normal_account,
                 &U256::MAX,
                 &mut substates.last_mut().unwrap(),
+                &mut NoopTracer,
                 Spec::new_spec_for_test().account_start_nonce
             )
             .unwrap(),
@@ -1340,6 +1352,7 @@ fn test_automatic_collateral_normal_account() {
                 &normal_account,
                 &U256::MAX,
                 &mut substates.last_mut().unwrap(),
+                &mut NoopTracer,
                 Spec::new_spec_for_test().account_start_nonce
             )
             .unwrap(),
@@ -1382,6 +1395,7 @@ fn test_automatic_collateral_normal_account() {
                 &normal_account,
                 &U256::MAX,
                 &mut substates.last_mut().unwrap(),
+                &mut NoopTracer,
                 Spec::new_spec_for_test().account_start_nonce
             )
             .unwrap(),
@@ -1426,6 +1440,7 @@ fn test_automatic_collateral_normal_account() {
                 &normal_account,
                 &U256::MAX,
                 &mut substates.last_mut().unwrap(),
+                &mut NoopTracer,
                 Spec::new_spec_for_test().account_start_nonce
             )
             .unwrap(),
@@ -1513,6 +1528,7 @@ fn test_automatic_collateral_contract_account() {
                 &contract_account,
                 &U256::MAX,
                 &mut substates.last_mut().unwrap(),
+                &mut NoopTracer,
                 Spec::new_spec_for_test().account_start_nonce
             )
             .unwrap(),
@@ -1550,6 +1566,7 @@ fn test_automatic_collateral_contract_account() {
                 &contract_account,
                 &U256::MAX,
                 &mut substates.last_mut().unwrap(),
+                &mut NoopTracer,
                 Spec::new_spec_for_test().account_start_nonce
             )
             .unwrap(),
@@ -1603,6 +1620,7 @@ fn test_automatic_collateral_contract_account() {
                 &contract_account,
                 &U256::MAX,
                 &mut substates.last_mut().unwrap(),
+                &mut NoopTracer,
                 Spec::new_spec_for_test().account_start_nonce
             )
             .unwrap(),
@@ -1649,6 +1667,7 @@ fn test_automatic_collateral_contract_account() {
                 &contract_account,
                 &U256::MAX,
                 &mut substates.last_mut().unwrap(),
+                &mut NoopTracer,
                 Spec::new_spec_for_test().account_start_nonce
             )
             .unwrap(),
@@ -1693,6 +1712,7 @@ fn test_automatic_collateral_contract_account() {
                 &contract_account,
                 &U256::MAX,
                 &mut substates.last_mut().unwrap(),
+                &mut NoopTracer,
                 Spec::new_spec_for_test().account_start_nonce
             )
             .unwrap(),
@@ -1738,6 +1758,7 @@ fn test_automatic_collateral_contract_account() {
                 &contract_account,
                 &U256::MAX,
                 &mut substates.last_mut().unwrap(),
+                &mut NoopTracer,
                 Spec::new_spec_for_test().account_start_nonce
             )
             .unwrap(),
