+++ core/state/src/lib.rs
@@ -10,6 +10,7 @@ pub mod state;
 pub(self) mod state_object_cache;
 pub mod state_trait;
 pub mod substate_trait;
+pub mod tracer;
 
 pub use state_trait::StateTrait;
 pub use substate_trait::{SubstateMngTrait, SubstateTrait};
