+++ core/state/src/state_trait.rs
@@ -17,13 +17,15 @@ pub trait StateTrait: CheckpointTrait {
     /// checked out. This function should only be called in post-processing
     /// of a transaction.
     fn settle_collateral_for_all(
-        &mut self, substate: &Self::Substate, account_start_nonce: U256,
+        &mut self, substate: &Self::Substate,
+        tracer: &mut dyn InternalTransferTracer, account_start_nonce: U256,
     ) -> DbResult<CollateralCheckResult>;
 
     // FIXME: add doc string.
     fn collect_and_settle_collateral(
         &mut self, original_sender: &Address, storage_limit: &U256,
-        substate: &mut Self::Substate, account_start_nonce: U256,
+        substate: &mut Self::Substate, tracer: &mut dyn InternalTransferTracer,
+        account_start_nonce: U256,
     ) -> DbResult<CollateralCheckResult>;
 
     // TODO: maybe we can find a better interface for doing the suicide
@@ -258,7 +260,7 @@ pub trait CheckpointTrait: StateOpsTrait {
 }
 
 use super::{CleanupMode, CollateralCheckResult};
-use crate::substate_trait::SubstateTrait;
+use crate::{substate_trait::SubstateTrait, tracer::InternalTransferTracer};
 use cfx_internal_common::{
     debug::ComputeEpochDebugRecord, StateRootWithAuxInfo,
 };
